[edr] [LINUXDAR-1278] change build process:  1. to build and run coverity from ci.  2. produce bullseye build and either run unit test or not  3. copy bullseye html results to output on artifactory  4. refactor upload so it can upload unittest results later  5. update to testBullseyeBuild.sh so as to run test on TAP template and on local Jenkins VM

+++ edr/build/bullseye/uploadResults.sh
@@ -14,29 +14,38 @@ function exitFailure()
 
 [[ -n ${BULLSEYE_DIR} ]] || exitFailure 1 "BULLSEYE_DIR not set"
 [[ -n ${COVFILE} ]] || exitFailure 2 "COVFILE not set"
-[[ -n ${COV_HTML_BASE} ]] || COV_HTML_BASE=sspl-unittest
-[[ -n ${htmldir} ]] || htmldir=output/coverage/${COV_HTML_BASE}
+[[ -n ${COV_HTML_BASE} ]] || COV_HTML_BASE=sspl-plugin-edr-combined
+[[ -n ${htmldir} ]] || htmldir=/opt/test/logs/coverage/${COV_HTML_BASE}
 
 PRIVATE_KEY=${BASE}/build/bullseye/private.key
 [[ -f ${PRIVATE_KEY} ]] || PRIVATE_KEY=build/bullseye/private.key
 [[ -f ${PRIVATE_KEY} ]] || exitFailure 3 "Unable to find private key for upload"
 
-echo "Exclusions:"
-covselect --list --no-banner --file "$COVFILE"
 
-$BULLSEYE_DIR/bin/covhtml \
-    --file "$COVFILE"     \
-    --srcdir /            \
-    --verbose             \
-    "$htmldir"            \
-    </dev/null            \
-    || exitFailure $FAILURE_BULLSEYE "Failed to generate bulleye html"
+if [[ -z ${UPLOAD_ONLY} ]]
+then
+  echo "Exclusions:"
+  $BULLSEYE_DIR/bin/covselect --list --no-banner --file "$COVFILE"
+
+  #Will produce an error if the source directory is not available but the output html is generated anyway
+  #When source is not avaible the output shows down to the function/member function level
+  $BULLSEYE_DIR/bin/covhtml \
+      --file "$COVFILE"     \
+      --srcdir /            \
+      --verbose             \
+      "$htmldir"            \
+      </dev/null            \
+      || exitFailure $FAILURE_BULLSEYE "Failed to generate bulleye html"
+fi
 
 chmod -R a+rX "$htmldir"
 
-## Ensure ssh won't complain about private key permissions:
-chmod 600 ${PRIVATE_KEY}
-rsync -va --rsh="ssh -i build/bullseye/private.key" --delete $htmldir \
-    upload@allegro.eng.sophos:public_html/bullseye/  \
-    </dev/null \
-    || exitFailure $FAILURE_BULLSEYE "Failed to upload bulleye html"
+if [[ ${BULLSEYE_UPLOAD} == 1 ]]
+then
+  ## Ensure ssh won't complain about private key permissions:
+  chmod 600 ${PRIVATE_KEY}
+  rsync -va --rsh="ssh -i build/bullseye/private.key" --delete $htmldir \
+      upload@allegro.eng.sophos:public_html/bullseye/  \
+      </dev/null \
+      || exitFailure $FAILURE_BULLSEYE "Failed to upload bulleye html"
+fi
